+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to the [Camunda Modeler](https://github.com/camunda/camunda-
 ___Note:__ Yet to be released changes appear here._
 
 * `FEAT`: add ability to copy diagram file path ([#6059](https://github.com/camunda/camunda-modeler/pull/6059))
+* `FIX`: surface the reason of instance start failed during task testing ([#6076](https://github.com/camunda/camunda-modeler/pull/6076))
+
+## 5.50.1
+
+* `FIX`: make `chrome_crashpad_handler` in Linux distribution executable ([#6080](https://github.com/camunda/camunda-modeler/pull/6080))
 
 ## 5.50.0